Move TaskContext class definition from Microsoft.Spark.Worker to Microsoft.Spark
